pyproject.toml
[project]
name = "gigatime"
version = "0.1.0"
description = "GigaTIME: virtual mIF prediction from H&E images"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.4",
    "numpy>=2.2",
    "pillow>=10.4",
    "albumentations>=1.4",
    "huggingface-hub>=0.24",
    "openslide-python",
    "openslide-bin",
    "boto3>=1.34",
    "owkin-abstra>=1.8.1",
    "owkin-sagemaker-toolbox",
    "owkin-tilingtoolv2>=11.16.1",
    "classic-algos",
]

[project.optional-dependencies]
train = [
    "pandas>=2.3",
    "scipy>=1.16",
    "scikit-learn>=1.5",
    "scikit-image>=0.25",
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "easydict>=1.13",
    "opencv-python-headless>=4.10",
]
notebooks = [
    "gigatime[train]",
    "jupyterlab>=4.2",
    "ipykernel>=6.29",
    "ipywidgets>=8.1",
    "matplotlib>=3.9",
]
dev = [
    "gigatime[notebooks]",
    "pytest>=8.0",
    "ruff>=0.5",
    "mypy>=1.10",
    "boto3-stubs[s3]>=1.34",
    "pre-commit>=3.7",
    "pre-commit-hooks>=4.6",
    "codespell>=2.3",
    "yamllint>=1.35",
]

[[tool.uv.index]]
name = "codeartifact"
url = "https://abstra-058264397262.d.codeartifact.eu-west-1.amazonaws.com/pypi/owkin-pypi/simple/"
explicit = true

[tool.uv.sources]
owkin-sagemaker-toolbox = { index = "codeartifact" }
owkin-abstra = { index = "codeartifact" }
owkin-tilingtoolv2 = { index = "codeartifact" }
classic-algos = { index = "codeartifact" }

[project.scripts]
gigatime-infer = "gigatime.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gigatime"]

[tool.mypy]
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
